+++ mayan/apps/metadata/api_views.py
@@ -15,7 +15,7 @@ from permissions import Permission
 from rest_api.filters import MayanObjectPermissionsFilter
 from rest_api.permissions import MayanPermission
 
-from .models import DocumentMetadata, MetadataType
+from .models import DocumentMetadata, DocumentTypeMetadataType, MetadataType
 from .permissions import (
     permission_metadata_document_add, permission_metadata_document_remove,
     permission_metadata_document_edit, permission_metadata_document_view,
@@ -23,8 +23,9 @@ from .permissions import (
     permission_metadata_type_edit, permission_metadata_type_view
 )
 from .serializers import (
-    DocumentMetadataSerializer, DocumentTypeNewMetadataTypeSerializer,
-    MetadataTypeSerializer, DocumentTypeMetadataTypeSerializer
+    DocumentMetadataSerializer, DocumentNewMetadataSerializer,
+    DocumentTypeNewMetadataTypeSerializer, MetadataTypeSerializer,
+    DocumentTypeMetadataTypeSerializer
 )
 
 
@@ -89,15 +90,14 @@ class APIMetadataTypeView(generics.RetrieveUpdateDestroyAPIView):
 
 class APIDocumentMetadataListView(generics.ListCreateAPIView):
     permission_classes = (MayanPermission,)
-    serializer_class = DocumentMetadataSerializer
 
     def get_document(self):
-        return get_object_or_404(Document, pk=self.kwargs['document_pk'])
+        return get_object_or_404(Document, pk=self.kwargs['pk'])
 
     def get_queryset(self):
         document = self.get_document()
 
-        if self.request == 'GET':
+        if self.request.method == 'GET':
             # Make sure the use has the permission to see the metadata for
             # this document
             try:
@@ -111,7 +111,7 @@ class APIDocumentMetadataListView(generics.ListCreateAPIView):
                 )
             else:
                 return document.metadata.all()
-        elif self.request == 'POST':
+        elif self.request.method == 'POST':
             # Make sure the use has the permission to add metadata to this
             # document
             try:
@@ -126,14 +126,21 @@ class APIDocumentMetadataListView(generics.ListCreateAPIView):
             else:
                 return document.metadata.all()
 
-    def pre_save(self, serializer):
-        serializer.document = self.get_document()
+    def get_serializer_class(self):
+        if self.request.method == 'GET':
+            return DocumentMetadataSerializer
+        elif self.request.method == 'POST':
+            return DocumentNewMetadataSerializer
 
     def get(self, *args, **kwargs):
         """
         Returns a list of selected document's metadata types and values.
         """
         return super(APIDocumentMetadataListView, self).get(*args, **kwargs)
+
+    def perform_create(self, serializer):
+        serializer.document = self.get_document()
+        serializer.save()
 
     def post(self, *args, **kwargs):
         """
@@ -265,10 +272,15 @@ class APIDocumentTypeMetadataTypeOptionalListView(generics.ListCreateAPIView):
             metadata_type = get_object_or_404(
                 MetadataType, pk=serializer.data['metadata_type_pk']
             )
-            document_type.metadata_type.add(
-                metadata_type, required=self.required_metadata
+            document_type_metadata_type = document_type.metadata.create(
+                metadata_type=metadata_type, required=self.required_metadata
             )
-            return Response(status=status.HTTP_201_CREATED)
+            return Response(
+                status=status.HTTP_201_CREATED,
+                data={
+                    'pk': document_type_metadata_type.pk
+                }
+            )
         else:
             return Response(status=status.HTTP_400_BAD_REQUEST)
 
@@ -291,18 +303,19 @@ class APIDocumentTypeMetadataTypeRequiredListView(APIDocumentTypeMetadataTypeOpt
         """
         return super(
             APIDocumentTypeMetadataTypeRequiredListView, self
-        ).get(*args, **kwargs)
+        ).post(request, *args, **kwargs)
 
 
-class APIDocumentTypeMetadataTypeRequiredView(views.APIView):
+class APIDocumentTypeMetadataTypeView(views.APIView):
     def delete(self, request, *args, **kwargs):
         """
         Remove a metadata type from a document type.
         """
 
-        document_type = get_object_or_404(
-            DocumentType, pk=self.kwargs['document_type_pk']
+        document_type_metadata_type = get_object_or_404(
+            DocumentTypeMetadataType, pk=self.kwargs['pk']
         )
+
         try:
             Permission.check_permissions(
                 self.request.user, (permission_document_type_edit,)
@@ -310,11 +323,8 @@ class APIDocumentTypeMetadataTypeRequiredView(views.APIView):
         except PermissionDenied:
             AccessControlList.objects.check_access(
                 permission_document_type_edit, self.request.user,
-                document_type
+                document_type_metadata_type.document_type
             )
 
-        metadata_type = get_object_or_404(
-            MetadataType, pk=self.kwargs['metadata_type_pk']
-        )
-        document_type.metadata_type.remove(metadata_type)
+        document_type_metadata_type.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)