+++ mayan/apps/metadata/serializers.py
@@ -19,6 +19,7 @@ class DocumentMetadataSerializer(serializers.ModelSerializer):
     class Meta:
         fields = ('document', 'id', 'metadata_type', 'value',)
         model = DocumentMetadata
+        read_only_fields = ('metadata_type',)
 
 
 class DocumentTypeMetadataTypeSerializer(serializers.ModelSerializer):
@@ -28,16 +29,34 @@ class DocumentTypeMetadataTypeSerializer(serializers.ModelSerializer):
 
 
 class DocumentNewMetadataSerializer(serializers.Serializer):
-    metadata_type = serializers.IntegerField(
-        help_text=_('Primary key of the metadata type to be added.')
+    metadata_type_pk = serializers.IntegerField(
+        help_text=_('Primary key of the metadata type to be added.'),
+        write_only=True
     )
+
+    metadata_type = MetadataTypeSerializer(read_only=True)
+
+    pk = serializers.IntegerField(
+        help_text=_('Primary key of the document metadata type.'),
+        read_only=True
+    )
+
     value = serializers.CharField(
         max_length=255,
         help_text=_('Value of the corresponding metadata type instance.')
     )
 
+    def create(self, validated_data):
+        metadata_type = MetadataType.objects.get(
+            pk=validated_data['metadata_type_pk']
+        )
+        instance = self.document.metadata.create(
+            metadata_type=metadata_type, value=validated_data['value']
+        )
+        return instance
+
 
 class DocumentTypeNewMetadataTypeSerializer(serializers.Serializer):
-    metadata_type = serializers.IntegerField(
+    metadata_type_pk = serializers.IntegerField(
         help_text=_('Primary key of the metadata type to be added.')
     )